SqlAgent/pyproject.toml
[tool.poetry]
name = "sqlagent"
version = "0.1.0"
description = ""
authors = ["Manimaran Ramakrishnan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.7"
langgraph = "^0.2.45"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
